Move to V6.4 and add catch if exists
