+++ src/Core/Migration/Migration1664541794AddIndexForTasks.php
@@ -1,0 +1,13 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Core\Migration;
+
+use Doctrine\DBAL\Connection;
+use Shopware\Core\Framework\Migration\MigrationStep;
+
+/**
+ * @deprecated tag:v6.4.0 - Will be deleted. Migrations are now namespaced by major version
+ */
+class Migration1664541794AddIndexForTasks extends \Shopware\Core\Migration\V6_4\Migration1664541794AddIndexForTasks
+{
+}
